+++ dotnet/Services/IKlaviyoAPI.cs
@@ -6,7 +6,7 @@ namespace Klaviyo.Services
 {
     public interface IKlaviyoAPI
     {
-        Task<KlaviyoEvent> BuildEvent(string orderId, string eventType, DateTimeOffset lastChange);
+        Task<KlaviyoEvent> BuildEvent(VtexOrder vtexOrder, string eventType, DateTimeOffset lastChange);
         Task<bool> ProcessNotification(HookNotification hookNotification);
         Task<bool> SendEvent(KlaviyoEvent klaviyoEvent);
     }